Stata: do&log files
a. Determine how to read both data sets into Stata and merge them together by the participant id SEQN.

b. Use logistic regression to estimate the relationship between age (in months) and the probability that an individual has a primary rather than a missing or permanent upper right 2nd bicuspid. You can recode permanent root fragments as permanent and drop individuals for whom this tooth was not assessed. Use the fitted model to estimate the ages at which 25, 50, and 75% of individuals lose their primary upper right 2nd bicuspid. Round these to the nearest month. Choose a range of representative age values with one year increments by taking the floor (in years) of the 25%-ile and the ceiling (in years) of the 75%-ile.

c. In the regression above, control for demographics in the following way:
        1. Add gender to the model and retain it if it improves the BIC.
        2. Create indicators for each race/ethnicity category using the largest as the reference and collapsing ‘Other Hispanic’ and ‘Other’. In order of group size in the sample, add each category retaining those that improve BIC.
        3. Add poverty income ratio to the model and retain it if it improves BIC.

d. Use the margins command to compute:
        1. Adjusted predctions at the mean (for other values) at each of the representative ages determined in part b.
        2. The marginal effects at the mean of any retained categorical variables at the same representative ages.
        3. The average marginal effect of any retained categorical varialbes at the representative ages.

    [5 points] Refit your final model from part c using svy and comment on the differences. Include a nicely formatted regression table and cite evidence to justify your comments.